requirements.txt
streamlit
streamlit-lottie
torch
plotly
pandas
numpy
scikit-learn
requests
PyPDF2
pillow
python-dotenv
nltk
regex
joblib
matplotlib
seaborn
docx2txt
spacy
pyyaml
tqdm
openai
jsonschema
beautifulsoup4
lxml